+++ spec/langchain/llm/ollama_spec.rb
@@ -3,7 +3,7 @@
 require "faraday"
 
 RSpec.describe Langchain::LLM::Ollama do
-  let(:subject) { described_class.new(url: "http://localhost:11434", default_options: {completion_model_name: "llama3", embeddings_model_name: "llama3"}) }
+  let(:subject) { described_class.new(url: "http://localhost:11434", default_options: {completion_model_name: "llama3.1", embeddings_model_name: "llama3.1"}) }
   let(:client) { subject.send(:client) }
 
   describe "#initialize" do
@@ -123,7 +123,7 @@ RSpec.describe Langchain::LLM::Ollama do
 
   describe "#default_dimensions" do
     it "returns size of llama3 embeddings" do
-      subject = described_class.new(url: "http://localhost:11434", default_options: {embeddings_model_name: "llama3"})
+      subject = described_class.new(url: "http://localhost:11434", default_options: {embeddings_model_name: "llama3.1"})
 
       expect(subject.default_dimensions).to eq(4_096)
     end