+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 ## [Unreleased]
+- [BREAKING] Langchain::LLM::OpenAI#chat() uses `gpt-4o-mini` by default instead of `gpt-3.5-turbo` previously.
+- [BREAKING] Assistant works with a number of open-source models via Ollama.
+- [BREAKING] Introduce new `Langchain::ToolDefinition` module to define tools. This replaces the previous reliance on subclassing from `Langchain::Tool::Base`.
+
+## [0.14.0] - 2024-07-12
+- Removed TokenLength validators
+- Assistant works with a Mistral LLM now
+- Assistant keeps track of tokens used
+- Misc fixes and improvements
 
 ## [0.13.5] - 2024-07-01
 - Add Milvus#remove_texts() method